+++ src/lib/index.ts
@@ -1,0 +1,3 @@
+export function isString(n: unknown) {
+  return typeof n === "string" || n instanceof String;
+}